pyproject.toml
[project]
name = "norwood-ai"
version = "0.1.0"
description = "Norwood scale classification API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "anthropic>=0.39.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "celery[redis]>=5.4.0",
    "boto3>=1.35.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "python-ulid>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "pillow>=12.0.0",
    "stripe>=8.0.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
